Open menu and exit always showing. Code refactor.
